Extract ForGuidIdentifierSwapper to base class
