+++ src/microservices/Microservices.IdentifierMapper/Execution/Swappers/TableLookupWithGuidFallbackSwapper.cs
@@ -43,13 +43,13 @@ namespace Microservices.IdentifierMapper.Execution.Swappers
             _tableSwapper.Setup(mappingTableOptions);
 
             var guidOptions = mappingTableOptions.Clone();
-            guidOptions.MappingTableName = GetGuidTableIfAny(guidOptions).GetFullyQualifiedName();
+            guidOptions.MappingTableName = GetMappingTableIfAny(guidOptions).GetFullyQualifiedName();
             guidOptions.ReplacementColumnName = GuidColumnName;
             _guidSwapper.Setup(guidOptions);
         }
 
         /// <summary>
-        /// Returns the main lookup table, for the temporary guid allocations use <see cref="GetGuidTableIfAny(IMappingTableOptions)"/>
+        /// Returns the main lookup table, for the temporary guid allocations use <see cref="GetMappingTableIfAny(IMappingTableOptions)"/>
         /// </summary>
         /// <returns></returns>
         public DiscoveredTable GetMappingTable(IMappingTableOptions options)
@@ -63,7 +63,7 @@ namespace Microservices.IdentifierMapper.Execution.Swappers
         /// </summary>
         /// <param name="options"></param>
         /// <returns></returns>
-        public override DiscoveredTable GetGuidTableIfAny(IMappingTableOptions options)
+        public override DiscoveredTable GetMappingTableIfAny(IMappingTableOptions options)
         {
             var mappingTable = options.Discover();
             var guidTableName = mappingTable.GetRuntimeName() + GuidTableSuffix;
